Adapt ImageType for BraveAutoplayBlockedImageModel
As clarified by @darkdh, the ImageType passed to the constructor of
ContentSettingSimpleImageModel to build BraveAutoplayBlockedImageModel
does not really matter since we will eventually set our own icon vector
so we can set anything other than ImageType::PLUGIN now that such value
has been dropped upstream.

With this in mind, we're picking MEDIASTREAM just to fullfill with the
constructor's signature, as we don't really have a better choice.

Chromium change:

https://source.chromium.org/chromium/chromium/src/+/3149332f5447509043835fbb7207ede107c738dd

commit 3149332f5447509043835fbb7207ede107c738dd
Author: Illia Klimov <elklm@google.com>
Date:   Tue Nov 24 16:53:03 2020 +0000

    Reduce PLUGINS call sites.

    This CL reduces call sites for ContentSettingsType::PLUGINS.

    Bug: 1149393